feat: add back navigation (#3386)
* chore: extract connect manually flow to individual component
