+++ packages/frontend/src/components/ProjectConnection/DbtSettingsForm.tsx
@@ -5,6 +5,7 @@ import {
 } from '@lightdash/common';
 import { FC, useMemo, useState } from 'react';
 import { useFormContext, useWatch } from 'react-hook-form';
+import { SelectedWarehouse } from '../../pages/CreateProject';
 import { useApp } from '../../providers/AppProvider';
 import FormSection from '../ReactHookForm/FormSection';
 import Input from '../ReactHookForm/Input';
@@ -16,7 +17,6 @@ import DbtCloudForm from './DbtForms/DbtCloudForm';
 import DbtLocalForm from './DbtForms/DbtLocalForm';
 import GithubForm from './DbtForms/GithubForm';
 import GitlabForm from './DbtForms/GitlabForm';
-import { SelectedWarehouse } from './ProjectConnectFlow/WareHouseConnectCard.tsx';
 import {
     AdvancedButton,
     AdvancedButtonWrapper,
